Split console methods
