+++ src/Codeception/Module/Laravel/InteractsWithConsole.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Codeception\Module\Laravel;
+
+use Illuminate\Contracts\Console\Kernel;
+use Symfony\Component\Console\Output\OutputInterface;
+
+trait InteractsWithConsole
+{
+    /**
+     * Call an Artisan command.
+     *
+     * ``` php
+     * <?php
+     * $I->callArtisan('command:name');
+     * $I->callArtisan('command:name', ['parameter' => 'value']);
+     * ```
+     * Use 3rd parameter to pass in custom `OutputInterface`
+     *
+     * @param string $command
+     * @param array $parameters
+     * @param OutputInterface|null $output
+     * @return string|void
+     */
+    public function callArtisan(string $command, $parameters = [], OutputInterface $output = null)
+    {
+        $console = $this->app->make(Kernel::class);
+        if (!$output) {
+            $console->call($command, $parameters);
+            $output = trim($console->output());
+            $this->debug($output);
+            return $output;
+        }
+
+        $console->call($command, $parameters, $output);
+    }
+}
